+++ react/typings/vtex.store-resources.d.ts
@@ -1,0 +1,3 @@
+declare module 'vtex.store-resources/Queries' {
+  export const session
+}