+++ react/typings/vtex.styleguide.d.ts
@@ -1,0 +1,17 @@
+declare module 'vtex.styleguide' {
+  export const ButtonWithIcon
+  export const withToast
+  export const Button
+  export const Input
+  export const Toggle
+  export const ActionMenu
+  export const IconOptionsDots
+  export const Checkbox
+  export const IconDelete
+  export const Spinner
+  export const IconPlusLines
+  export const IconCaretLeft
+  export const IconClose
+  export const IconVisibilityOff
+  export const IconVisibilityOn
+}